MKT-LOGISTIC-ON-DEMAND/backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
bcrypt==5.0.0
click==8.3.1
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.115.6
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
passlib==1.7.4
psycopg==3.2.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.12.5
pydantic-settings==2.6.1
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
sqlmodel==0.0.22
starlette==0.41.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
